+++ redfishAPI.py
@@ -75,9 +75,7 @@ if __name__ == '__main__':
 
     # Before existing, stop the listener service if running
     if redfishConfig.listener != None:
-        redfishConfig.listener.stop = True
-        redfishConfig.listener.join()
-        redfishConfig.listener = None
+        redfishConfig.listener.shutdown()
 
     # Before existing, delete any current active session
     sessionId = Label.decode(config.sessionIdVariable)